Encapsulate types, finish translation function
